# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Login Into Admin Account",
                "operationId": "71828f836af3373bc948cade5d8ba326",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/admin/create": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create A New Admin Account",
                "operationId": "727a08275c6f285499b4ab6bbe4f539b",
                "requestBody": {
                    "description": "Supply Admin Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserProperties"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/admin/user-listing": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "View All Users",
                "operationId": "6409b580dc46970289ea9d2b3177b632",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page Number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Item Per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort By",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Direction of sort",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "User Address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User Phone",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Created At",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "marketing",
                        "in": "query",
                        "description": "User Is marketing",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/user-edit/{uuid}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update User Account",
                "operationId": "072d22e651e6b7b5b867ffe0d59339d4",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of User",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "uuiuytyytytj5656jk-jnnknnkbkjnk-nghn6n565"
                    }
                ],
                "requestBody": {
                    "description": "Supply Admin Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserProperties"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/user-delete/{uuid}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete A User Account",
                "operationId": "346c861a088343f73d85add500fcf798",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of User",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "uuiuytyytytj5656jk-jnnknnkbkjnk-nghn6n565"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateUserProperties": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "format": "first_name",
                        "example": "first_name"
                    },
                    "last_name": {
                        "type": "string",
                        "format": "last_name",
                        "example": "last_name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@email.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "avatar": {
                        "type": "string",
                        "format": "avatar",
                        "example": "avatar"
                    },
                    "address": {
                        "type": "string",
                        "format": "address",
                        "example": "address"
                    },
                    "phone_number": {
                        "type": "string",
                        "format": "phone_number",
                        "example": "phone_number"
                    },
                    "is_marketing": {
                        "type": "string",
                        "format": "is_marketing",
                        "example": "is_marketing"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CreateUserProperties"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "scheme": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}